Add locale loading, extract ResoniteMonkey hook from ModLoader class, update version
