+++ MonkeyLoader.GamePacks.ResoniteModLoader/ModConfigurationKey.cs
@@ -158,7 +158,7 @@ namespace ResoniteModLoader
         {
             if (string.IsNullOrWhiteSpace(name))
             {
-                ModLoader.Logger.Warn(() => $"ModConfigurationKey with description [{description}] has null or whitespace name - using Spacer name!");
+                ModLoaderHook.Logger.Warn(() => $"ModConfigurationKey with description [{description}] has null or whitespace name - using Spacer name!");
 
                 Key = new($"Spacer-{name?.GetHashCode() ?? description?.GetHashCode() ?? _replacementCounter++}",
                     description, computeDefault, internalAccessOnly, valueValidator);
